api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
pandas==2.2.2
requests==2.32.3
redis==5.0.6
python-jose[cryptography]==3.3.0
email-validator==2.1.1
jinja2==3.1.4
sqlglot==23.7.0
httpx==0.27.0
python-dotenv==1.0.1
nemoguardrails==0.15.0
